feat: add support for monorepos in open-source projects
Some relevant projects are sub-projects of larger projects,
and are accessible via sub-directories of monorepos, rather than
the root of a particular repository.

In these instances, we want to be able to link directly to those
subdirectories, and (optionally) provide an override for the
description, rather than taking the info from the monorepo.