docs(contribute): Move data note to list
I assiduously checked the list, twice, to make sure I was following it. However, I missed the bold note. I think that adding it to the list, explicitely, but keeping the bold will ensure that that this will not happen to other potential contributors. Something about being the same color and weight as the header made it easy for my eyes to glide over.